requirements.txt
langchain
openai
python-dotenv
langchain_openai